requirements.txt
annotated-types==0.7.0
lxml==6.0.2
pydantic==2.13.3
pydantic-core==2.46.3
python-docx==1.2.0
pyyaml==6.0.3
typing-extensions==4.15.0
typing-inspection==0.4.2
python-doxc
